requirements.txt
langchain
langgraph
langchain-core
langchain-google-genai
google-genai
pydantic
langchain-text-splitters
python-dotenv
flask
flask-cors
